cloudflare are picky when they man-in-the-middle u
Let's Encrypt certificate generation and renwal has trouble when there is a man-in-the-middle SSL cert thing going on,
as in the case of a CDN protected website that serves SSL on behalf of the origin website.

See
https://community.letsencrypt.org/t/options-for-renewing-certificates-with-cloudflare-proxying-enabled/35804/3
for a similar problem.

Rather than change the SSL re-generation container (too much work, can wait for Jan), we can just serve /static and /media over HTTP.